Bugfix: bug #21784
Symbols from gresym are always printed at current font size.  Implementation is perhaps a bit clunky, as it reloads the font when it detects a mismatch between the current font size and the font size of gresym as currently loaded.  This means that switching back and forth between two different font sizes will cause the font to be reloaded over and over again.

Processing would be more efficient if each time gresym was called for at a new size, it was loaded under a new name and the list of names (and their sizes) currently loaded was tracked.  In this fashion, when a size is reused, the existing loaded version at that size could be used.  However, I don't know how that might be implemented.