Script switch
Switched over the fontforge script to a python script.
Further modifications to Makefile to reflect the change in script and complete the instructions for creating the fonts.
Updated INSTALL to reflect changes.  This may need further updating as it indicates that squarize.py should generate a fontforge script which in turn generates the font.  However, in my testing it appeared that squarize.py generated the font directly.