Addressing #41
A new dimension is added which allows the user to specify the width of the box for the initial.  The initial will be centered in a box of the given width.  If the no width is given (i.e. it is 0) then the width of the box is what is necessary to contain the text within it.  Right now this doesn't work well with em and ex units because they are used at the current font where the length is set in the document, not the font of the initial.  Trying to switch to the initial font via \greinitialformat doesn't work (and appears to break the use of \gremanualinitialwidth).
Also fixes bugs in the behavior of \grebeforeinitialshift and \greafterinitialshift (these dimensions weren't rescaling appropriately with \grefactor).