Changing user settable distances to string representation
Distance setting and manipulation functions in gregoriotex-spaces.tex adapted to expect and store distances as strings.  This allows distances to be set in em and ex units and have that be respected at time of use.
Distances now all have a flag which tells whether they should scale when \grefactor changes
Distances in gsp-default.tex are set with \gresetdim
Create a temporary distance measure before \kern, \hskp, and \grehskip calls so that they are passed a distance, not a string
Also applies fix-66 to this branch (can't do this via merge because of conflicting changes)

** Still broken, but I don't know why.  Need to get help.