Un/ReLocalizing variables
Register allocation is an inherently global process in TeX and each call of \new… reserves a register (even if it's the same name as before).  This means that the sprinkling of \new… commands I did in previous commits is a bad idea which might lead to a large document running out of registers.  This commit removes those commands where possible.  To prevent this from further holding up the release of the new spacing interface, further refinement of the localization of variables will come later.