Documentation and Bugfixing
Relevant documentation for spaces modifications added.
Added a \gre@debug function which will print to the log when \ifdebug is true.  The flag can be set to true manually (\debugtrue) or by calling the LaTeX package with the debug option (new option).
Squashed bug that led to glue values getting printed in some scores. (#75)