Add new macros to force gregoriotex to recompile scores.
`\forcegabccompile` - By using this macro in the main tex file the user,
can force gregoriotex to recompile scores in the document. Useful for
when there are changes/bug fixes to gregoriotex and the API VERSION is
not changed.

`\normalgabccompile` - reverses `\forcegabccompile`. Gregoriotex will
return to its automatic score compiling behavior.