Corrections to CHANGELOG
Fixed some spelling errors.
Removed mention of bugs which were introduced by the work on releases 2.4.3 and 3.0.0
Eliminated complicated explanation of `\includescore` changes in favor of reference to UserManual.  Still need to fill in the section number once it has been updated.
Noted change of distance names and added reference to UserManual.  Will need to fill in section number once it has been updated.
Eliminated extra lines in 2.4.3
Removed mention of bug fix for deprecated code in autoconfig.ac as most users won't interact with that file.
Moved entry for addition of CHANGELOG from 2.4.3 to 3.0.0