Changed version test to only fail on mismatched major versions.
If a score declares that is it for version `x.y.z` the score will only
need to be recompiled if the major version of gregorio changes,
i.e. `x+1.y.z`.

This does not effect the `[autocompile]` options. When `autocompile` is
`true` a new score will be compiled for any version change.