fix VersionUpdate not idempotent
With semver compliant version numbers, the VersionUpdate script was not taking
into account the "-beta" in "3.0.0-beta" while reading the version in the files
to modify. Thus it was adding "-beta" all the time. Now the regexp is more
semver compliant, and so running the script several times is ok.