Bug fix for #146
Fixes problem with complete score not compiling.  There was a typo in the distance names for `highchoralsignshift` and `lowchoralsignshift` in the algorithm that decided whether they were dimen or skips (resulting in them getting classified wrong).
Added in the checks against `\gre@scale@…` when changing `\grefactor`.  Not sure where this code went, because I know I've done it before.  Requests to not scale distances should now be respected.
Moved a couple of files from `examples` to `tests`.  They will eventually need to migrate to the gregorio-project/gregorio-test, but the state of my internet connection makes even getting this fix up difficult.