Continued work on Mac installer
Created the contrib raw package.  This will install the contents of the contrib folder in User/Shared/Gregorio.  It is set to be relocatable.
Made some corrections to the base package project and the creation script to make use of a build folder within macosx rather than macosx itself.
Changed `.gitignore` settings to ignore the build folder as well as built packages.