Continued work on Mac installer.
Upgraded `gregorio-base` to be the Gregorio distribution package and started adding the presentation elements.
Created a installer project to wrap the uninstall script.  This project has an empty payload and just runs the uninstall script as a post-flight script.
Small edit to `create_package.sh` so that working directory is correct when running `build.sh` etc.