Adding Compiled Documentation
Modifications to `create_package.sh` so that the `GregorioRef.pdf` is produced
Modified `gregorio-doc.pkgproj` to install `GregorioRef-##.pdf` (note, does not account for the fact that the version number is in the file name and will change with each build).
Modification to `.gitignore` so that the `_minited-…` folder will be ignored, regardless of what version of GregorioRef has been compiled.