Fixed clean target in doc folder
Clean target in doc folder wasn't working because of the jobname option.  With that option the `latexmk -C` has to be invoked with the same jobname and target file in order to work.
Also added a remove statement to get rid of the `_minted*` directory that's created as part of building `GregorioRef-###.pdf`