Fix #207
Changed timing of `\gre@computespaces`.  Now instead of calling it whenever we change a distance, we call it once at the beginning of each score.  GregorioTeX will now respect the document width on a score by score basis as a result.
Also removed some redundant code from `\setgrefactor` that was already being done as part of `\begingregorioscore`.