Restructured installer directories
Moved the dialog files from `build` to `Resources`, this should make it easier to make the switch to the Xcode tools in the future (based on what I've discovered so far)
Rather than cherry pick the files for the extras out of the repository in `Gregorio.pkgproj`, we now copy them to build/Gregorio with `create_package.sh`.  This enables Packages to just include that folder, rather than having to include files on an individual basis.
Also simplified inclusion of the main project files using similar considerations.  This greatly reduces the complexity of the file.
For safety sake, the GregorioTeX temporary location is now `/tmp/gregorio` (instead of just `/tmp`)  This makes clean up easier and is less likely to interfere with anything else a user might have in `/tmp`.