Debugging the installer scripts
Turns out that the installer scripts run in a restricted shell environment that doesn't have the full `PATH` available to it.  Only the basic `PATH=/usr/bin:/bin:/usr/sbin:/sbin` is available, which excludes the TeX binaries in `/usr/texbin`.  I thus had to hard code the path to `kpsewhich` in order to get the scripts to work.  The installers will build and run now, but only on systems where the TeX binaries are in `/usr/texbin`.
Also some general housekeeping in the various files.