Getting the new documentation name into Windows installer
Since the name of the documentation pdf changed, the ISS file had to be made aware of the new file name.
This also required a new rule set in `VersionManager.py` that marks the filename for updating (since it contains the version number).  Unfortunately, ISS doesn't allow inline comments, so this has to be handled by putting the comment on the line before and then modifying the next line.