Bars with something under them were appearing too close to previous element
I found a commented out kern statement which when uncommented increases the spacing in the reported MWE.  However, checking the log, this line has been commented out since its introduction back in 7e2bfa4 (which was in 2008).  As a result, I'm not sure why this bug would not have appeared in earlier versions.  My only guess is that there were some missing `%` at the end of lines which resulted in the addition of uncontrolled whitespace and so while the line was added because theoretically it was needed, it was commented out because visually it resulted in too much whitespace.  As part of my spacing changes, I did a whole bunch of work to make sure that most, if not all, of the lines in the package terminated with a `%`, so that might have caused the problem to show up.  Unfortunately, my progressive testing seems to indicate that this bug shows up about the same time as the glue leaks, making it difficult to tell exactly what was going on at that time because I can't get a clean output.
For no discernable reason, `\GreSyllablenotes` has the wrong width for bar lines.  To get around this I take advantage of the fact that `\grewritebar` sets `lastglyphwidth` as the last step of its process and sets it correctly.  I thus make use of `lastglyphwidth` instead of `\wd\GreSyllablenotes` in `\grebarsyllable`.
Added the new english centering code in `\grebarsyllable`.

### Not ready to merge.