Added comment in spaces et al.
Added a comment in the source code indicating that `\gresetdim` stores distances as strings and indicating why.
Added functionality to `\grechangedim` so that it checks to make sure the distance exists before changing it.  Previously it would have happily created a new distance if passed an unexpected name.  This should make it easier to catch typos when manipulating distances.