Upgrading functionality of debug
`\gre@debug` becomes `\gre@debugmsg` in code.
The message function can now take two arguments.  The first, optional, is a keyword, the second, required, is the message.  If the keyword appears in the list of keywords stored in `\gre@debug` then the message is printed to the log.  If no keyword is given (or keyword is `all`) then debug message will always be printed to the log when debug is active.
`\gre@debug` is created via the debug option by the `kvoptions` package.  Syntax is `debug` (print all debug messages) `debug=keyword` (print only those messages associated with keyword, or `debug={list,of,keywords}` (print those messages associated with any of the keywords in the list).
As demonstration, all `ifdim` debug messages are now associated with that keyword.