Choosing a better default initial size
Instead of using `huge` and `Huge` in LaTeX, we define default initial sizes of 40pt and 80pt (for normal and big initials respectively).  If fonts are infinitely scalable, then this makes the appearance of the defaults nicer.  If fonts are not infinitely scalable, then LaTeX will automatically use the closest available size and put warnings in the log file.