Changes to font commands.
Font modification command definitions which are specific to PlainTeX now appear in `gregoriotex.tex` instead of `gregoriotex-main.tex`.  This way PlainTeX definitions are only loaded when using PlainTeX.
Additionally the commands for bold and colored have been renamed according to the new convention.  The others will get renamed as they come up on the list.