Completes Gregorio controls.
Added a couple of functions I missed on my first pass to the documentation.
All functions that were CamelCase, but which were not Gregorio controls have been renamed.  I made a basic attempt to follow the naming conventions, but expect that these functions may be renamed again when I deal more explicitly with the class they fall into.  This required a conversion of some flags from macros which expanded to `0` or `1` to booleans.  User functions renamed in this manner have deprecation warnings.