Unifying the way formats are changed
The commands for formatting different elements of the score are unified into a single mechanism.  This mechanism uses environments in LaTeX and matching `\greenvironment` and `\endgreenvironment` commands in PlainTeX.
The unified mechanism allows for a single command, `\grechangeformat` to change them all.
The formatting commands still exist, for backwards compatibility purposes, but this bloats the code.  Unfortunately I can't include deprecation warnings for these since the proper usage was to redefine these commands, something which would eliminate any deprecation warning I could put in.  At some point we'll want to remove these artifacts.