Support for arbitrary number of lines in the annotation.
Annotation functions are consolidated into a single function `\greannotation` which can, through multiple calls, build up an annotation with an arbitrary number of lines.
There are now two distances associated with the annotation `annotationseparation` (a renaming of `aboveinitialseparation` for consistency) and `annotationraise`.  The first controls the distance between lines in the annotation, the second raises or lowers the annotation from it's default position (baseline of the first line aligned with the top line of the staff).  This second distance should be more transparent than the old method of measuring the height of a second argument to the annotation functions and more consistent with how other spaces are manipulated.