Fix neume font not loading at correct size.
`\gresetgregoriofont` was pulling double duty, both selecting the font and loading it.  As a result, pulling it from `\GreBeginScore` resulted in the font not loading at the correct size when the `refactor` was changed.  I fixed this by splitting out the loading code into it's own function which both `\gresetgregoriofont` and `\setgrefactor` can call to trigger a reloading of the font when either the choice of font or the size changes.