Removed hidden dependency
`\gresetstylefont` had a hidden dependency in that it required the original `\gresetgregoriofont` to be called after it.  The recent changes to that function meant that the wasn't always happening anymore.  I copied the line that was needed from `\gre@loadgregoriofont` to `\gresetstylefont` so that the calls of these two functions are independent of one another and can occur in any order or singly.