Renaming colored lines and updating GregorioRef
Putting a score within a the center environment was throwing the annotations out of position on the title page.  I've reworked the title page to achieve the same look, but without score being within the center environment.
I've also done some cleaning up of the GregorioRef files: updating the outdated commands and fixing typos.