Patching PlainTeX for message format
`\MessageBreak` and `\protect` are defined in LaTeX but not in PlainTeX.  As a result the deprecation warnings were causing PlainTeX to throw an error and fail to compile.  This patches `\greerror`, `\gre@warning`, and `\gre@bug` in PlainTeX so that it will within those messages `\MessageBreak` and `\protect` work as expected.