Adding Deprecation Messages to old formatting commands
This branch fixes two problems:
 1.  In code which redefined the old formatting commands to change the style of text elements, no deprecation messages where raised to alert the user to the new syntax.  These changes result in those messages being raised.
 2. If a document mixed the old and the new formatting commands on the same element, both were applied, which could lead to unexpected results.  This was especially apparent for above line text, where the default (small italics) was defined in both syntaxes.  This change makes it so that only the old syntax is applied (with the appropriate deprecation message being raised).

This commit applies the necessary changes for all the formats (translation, above lines text, staff lines, and initials) not changed in the earlier commits.
All lines affected by these changes should now be labeled with `DEPRECATED` or `keep this line`, as appropriate, to help with making the appropriate changes in future versions.