Adding option to have an illuminated initial
The `\greillumination` command allows the user to put an arbitrary object (such as a graphics file) into the initial box.  This allows the user to break out of the limitations of installed fonts to incorporate much more elaborate initials.
Note: if the deprecated `\greinitialformat` or `\grebiginitalformat` is used, the illumination will be ignored.
A new test which illustrates this feature (tex-output/illumination/illumination.tex) has been added to the test suite.