Distinguishing end of line shifts from beginning of line shifts
End of line shifts (`eolshift`) are for allowing the lyric text to stretch under the custos
Beginning of line shifts (`bolshift`) are to align the lines on the notes rather than the text
The code previously conflated these two and used the same flag to determine whether to apply each one.  I've distinguished the two here, giving them each their own flag (both default to true) and user command for changing the behavior.