Fixing end of line shifts so that they are applied
The way the code was written, end of line shifts were never being applied, this rewrite gets them to have an effect.
End of line shifts, in this implementation, **prevent** the lyrics of the last syllable from stretching under (and past) the custos at the end of a manually broken line.  This reverses the sense of what the documentation formerly said about `eolshifts`, but it matches more closely with the actual implementation.
End of line shifts are only active if the following 4 conditions are met:
  1. The user wants them active
  2. We are manually ending a line which is justified
  3. There is a custos at the end of the line
  4. The lyrics for the last syllable of the line are longer than their notes