End of line shifts now happen after every syllable
Behavior of end of line shifts regarding activation reverts so that active *allows* lyrics under custos
Renamed `@compute@` functions to `@calculate@` for consistency
Added enddifference shift to manual line breaks because where it occurs in \GreSyllable ends up being after the line break when a manual line break is invoked.  This is the reason why it looked like eolshifts were working in reverse.