Making sure string distances are terminated
One of the problems with the old `eolshift` calculation was that string distances within the calculation weren't properly terminated by a `\relax`.  Without that `\relax` TeX's length parser will keep looking ahead until if finds something which can't possibly expand to a length, thus potentially absorbing some of the subsequent code.
Using a sed script, I've added that terminal `\relax` to most (if not all) instances of a distance which ends a line.  I'm fairly certain that my sed script was general enough to catch all the string distances (the ones defined in gsp-default.tex, and the ones written out explicitly in the code) and probably caught a few of the distance registers (which I don't think technically need it) as well.