Initial alterations are ignored when aligning the first note of each line (bolshift)
I've added an additional component (`bolextra`) to the `bolshift` which allows an initial alteration to be ignored when aligning the first notes of each line.  This works with the current setting for `minimalspaceatlinebeginning` but there are a couple of problems:
  1. if `minimalspaceatlinebeginning` is increased the lines which start with an alteration do not move
  2. if the initial alteration is a sharp, it overlaps with the clef.  i.e. there is not enough space for the alteration
I have also noticed that `minimalspaceatlinebeginning` has an upper limit on what is respected.  This makes me think that `bolshift` isn't being calculated correctly (at least in this aspect) and may be the source of the above two problems.