Preventing conflicts which arise from trying to load both gregoriosyms and gregoriotex packages.
If the order is gregoriosyms -> gregoriotex, then an error is raised instructing the user to remove the loading of gregoriosyms.
If the order is gregoriotex -> gregoriosyms, then the loading of gregoriosyms is aborted silently to prevent conflicting definitions.  We do not raise an error in this case because we've already loaded everything in gregoriosyms by loading gregoriotex and so simply aborting the loading of gregoriosyms allows the compilation of the document to proceed normally.