Automate the insertion of the version number into CHANGELOG.md
VersionManager will now automatically insert the appropriate version number and date into CHANGELOG, effectively starting a new Unreleased block
I also noticed that when inserting the date on a PARSE_VERSION_DATE line (i.e. in the documentation) the space between the comma and the date was being lost if a single digit date was replaced by a double digit one.  I just modified the substitution rule a bit to prevent this.