Making Mac installer scripts El Capitan compliant
Because `/usr` cannot be written to on El Capitan, MacTeX cannot create `/usr/texbin` the way it used to.  Instead it creates `/Library/TeX/texbin`.  As a result, when looking for `kpsewhich` we need to look in that location too should the earlier attempts to find it fail.
Note: We leave the search of `/usr/texbin` in place for older installations.