Efficiency improvements to create_package.sh
Made it possible to skip the clean-up step at the end of the package build process.   I copied the rc file model from gregorio-test for this setting so that persistent personal preferences can be set.  The option `-c` toggles the state.
create_package.sh now will now check for the existence of the target directory for the different pieces of the installer before making them.  If it finds the target directory, it assumes that piece is already made and skips it.
This allows rapid testing of modifications to the installer without going through a full build each time.