gregoriotex.lua: overhaul access to Luatex fonts tables
Related to

    https://github.com/gregorio-project/gregorio/issues/717
    https://github.com/lualatex/luaotfload/issues/308

The font table was indexed directly which is unsafe when using
Luaotfload. Instead, font.getfont() should be used which is overridden
by the fontloader. Due to caching, those accesses are faster too than
indexing the font table directly as well as using the builtin
font.getfont() API.

In order to facilitate future extensions to the code, a number of
frequent lookups have been implemented as accessor functions.

       ----------------------------------------------------
        expecting font id       expecting name
       ----------------------------------------------------
        get_font_by_name()      get_score_font_id()
        get_font_resources()    get_score_font_resources()
       ----------------------------------------------------

The *score_font* varieties look up the string in the ``score_font``
table first. Since in the majority of cases font accesses are followed
directly by a lookup of the ``.resources`` subtable, this is covered by
the *resources* variants.