Revising TeXShop configuration script
This adds `gabc` to the list of file formats which TeXShop can save when saving a new file (previously, you had to save it as `.tex` and then change the extension in Finder)
I've also added `gaux` and `gtex` to the list of files removed by the "Trash Aux Files" command.  Note that since the `gtex` files generally do not have the same basename as the project file, they will only be deleted when "Trash Aux Files" is in aggressive mode.