Rewrite of placement algorithm
I wasn't satisfied with the positions of the elements in the previous algorithm and the fact that there were two algorithms (one for text first and one for notes first) was only making it harder to fix.  This revised algorithm has better placement and there's only one to fine tune should it not be totally correct.