Missing spaces around bars
After double checking `\gre@writebar` I realized that if the syllable contains a bar, and only a bar, then the box containing the bar line does *not* include the necessary space around the bar.  I've thus borrowed code from `\gre@caluclate@syllablefinalskip` to determine the appropriate spacing around a bar in my new algorithm.
This necessitated moving the calculation of `\gre@count@temp@one` to earlier in the process.