Add uninstall script to installer
This script will be run as the first step of installation (using the Unistaller which is placed in the installation directory).  It should remove GregorioTeX files which were copied into the texmf tree and TeX bin directory as part of the installation.
Instead of having separate scripts for MiKTeX and TeXLive, this combined script relies on checking `TEXMFDIST` for "miktex" or "texlive" to determine which distribution is present and act accordingly.