Consolidation of install script
Instead of having separate install scripts designed to setup TeXLive and MiKTeX, I've combined them into a single script which can setup either distribution.
Working on this revealed a couple of bugs in the uninstall script which have also been fixed.
Finally, I added the `runascurrentuser` flag to the Run and UninstallRun entries in the iss file to make sure that these scripts launch with the same privileges as the installer itself.