Add clarity
I was using "offset" for both the distance between the center of the bar and the center of the text and for the extra shifts applied to keep the former under the limits.  This was a bit confusing, so I've clarified the comments and messages so that the difference is clearer.  "Offset" is the distance from the center of the bar to the center of the text.  "Shift" is the extra move (away from centering) that an element gets to keep the size of the offset under the limits.
I've also clarified the messages about required and allocated space for the text and the bar line and switch from reusing `\gre@skip@temp@three` for calculating the required space for the notes to using `\gre@skip@temp@four`.  This allows me to hold on to the space required by the text (in `\gre@skip@temp@three`) and by the notes (`\gre@skip@temp@four`) and reuse these values in later calculations (in particular the ones I'm working on for collision avoidance).