Reorder printing commands
Instead of printing the bar first, we print the text first and then the bar to bring things more into line with the old algorithm and `\GreSyllable`
We also switch to using `\gre@hskip` for the last move at the end of the bar syllable, mirroring what was implemented in the previous commit for the start of the syllable.  This should balance the stretching around a bar.