Modifications to Windows batch file
These modifications are somewhat fragile because they depend on me knowing what the results of `findstr` are supposed to look like and trimming off the known junk from the version number instead of searching for and extracting the version number itself.  I do things this way because Windows does not have any native string operations that can easily extract a substring which matches a regex pattern.  I'm looking into the possibility of using mtx-grep (which is packaged with TeXLive automatically), but am not sure that the necessary grep options (particularly -o) are available.