Prevent multiple copies of same addition
When one upgrades TeXShop the gabc file extension may get dropped from the list of recognized file extensions (since that information is stored in the application’s internal plist).  This may prompt the user to rerun this script.  However, the extensions added to the defaults are preserved through updates, so we don’t need to add those again.  We thus check for the presence of each extension in each list and only add those that are missing.