Limit files installed
In order to make things more consistent, the source files for fonts not included in the default distribution have been stripped out of the distribution tarsal.
`install-gtex.sh` has also been updated to so that it installs whatever is present, regardless of whether that’s just the distribution font, or a full build with all the extra fonts.