Install script for supplemental fonts
This lua script will install the supplemental fonts, both compiled and source files.  It assumes that the supplemental fonts to be installed are in the same directory as itself (and thus works from the repository once the fonts have been built).  The target directory is determined by passing an (optional) argument to the script.

 * If the argument is `auto` (or not given) then the script will find greciliae and install the fonts in the same location (if greciliae is not found, installation is aborted with an error message).
 * If the argument is `system` or `user`, then the fonts will be installed to `TEXMFLOCAL` or `TEXMFHOME` respectively (same behavior as install-gtex.sh).
 * Other argument values are a target root texmf directory for the installation.

In the latter two cases the resulting installation will follow the normal texmf tree structure.