Remove TeXShop auto-configure script
As it turns out, the script was causing TeXShop to misbehave, even to the point of not opening.  After some emailing back and forth with Richard Koch (the maker of TeXShop) it’s been worked out that support opening and editing Gregorio related files will appear built-in to the next release of TeXShop itself.  What I’ve done, therefore, is removed the problematic portions of the configuration script.  All it really does now is copy the engine into the appropriate directory.