Eliminate redundancy
I noticed that there was code in gregoriotex-main.tex which had to be duplicated in gregoriosyms.sty in order for that functionality to work properly in the symbols only style.  However, some recent changes in gregoriotex-main.tex weren’t reflected in the code in greogiosyms.sty resulting in an inconsistency.  Rather than simply fix that inconsistency, I’ve put the code that is supposed to be common to both in a separate file and included that file in both locations.  This way there’s only one place where the changes have to be made and they’ll be reflected in both uses.