Set trace to not expand arguments
Some arguments do not take well to expansion attempts before being written to the log and by not expanding the arguments what appears in the trace log is the function as it was called in the code, making it easier to find the right code location for an error (which is the whole point of this functionality).  Therefore, we prevent `\gre@trace` from expanding its arguments.