Script for uploading package to CTAN
This adds a script which makes use of ctan-o-mat (a script which comes with TeXLive) to upload the package to CTAN.  Naturally the CTAN package file must be built before it can be uploaded, so we check for it before trying to upload, but do not build it if it is not there.
The script attempts to scrape as much information as possible from the repository in creating the upload, but the user should always supply the text of the release message when invoking the script (as the first argument).  This can be skipped, but it is not recommended so the script asks for confirmation before letting that happen.
Finally, if it is needed, a second message can be added which is directed to the CTAN upload managers.  This is always the second argument of the script and can be left blank without a problem.