Reverse #1363
This reverses the work around we introduced to avoid a bug in LuaTeX in TL2017.  Said bug has been fixed in TL2018 (indeed, it was fixed shortly after TL2017 was released, but as it was a binary change to an engine, the fix was never introduced in the TL2017 update stream).

Obviously this change will break the package with TL2017.  Since the bug wasn't present in earlier versions of LuaTeX, compatibility with earlier versions of TL2016 will not be affected.  It will thus introduce a hole in the supported versions.  This has been documented in UPGRADE.md.