Relocate gtex and glog files
gregorio has trouble writing gtex files to arbitrary locations when security is `openout_any = p` (a problem we've encountered before in relation to snippets).  This solves the problem, but only for files found with kpse, by relocating the gtex and glog files to the project directory, rather than the directory where the gabc file is located.