Clean-up failed install
The uninstall script stub is installed fairly early in the process, before many of the installed files are put into place (it's populated as the files are installed).  Since the presence of this script is considered a block to future installations (barring the use of the AUTO_UNINSTALL variable) this makes life difficult when the installation fails part-way through.  Since a partial installation won't work anyway, this change makes use of the uninstall script stub to remove the partial install and then removes the stub itself, clearing the way for a clean installation after the problem causing the failure has been fixed.