Internalize gsp-default.tex
gsp-default.tex is turned into an internal file, not one meant for user access.  To do this we rename `\grecreatedim` to `\gre@createdim` (changing it to the internal namespace) and add a version check to the file.
We also change how scaling is applied when loading a space configuration file.  Instead of rescaling all distances when the space configuration file is loaded, we modify `\grechangedim` so that it individually checks to see if the distance being changed needs to be rescaled.  This allows user `gsp-*.tex` files to be partial, only defining the spacings they want to customize.