Space warning
`kpse.find_file` can find files along paths which have a space in their name, but `\input` cannot load them.  This doesn't matter if the file found in this manner is a gabc file (as its gtex file will be put in the current working directory, which `\input` has no problem with) but if a gtex file is found like this, then TeX will complain about being unable to find the file when it tries to input it.  To help indicate the problem, we check for a space in the path and issue a warning if we see one.  We only warn because gabc files will still work.