Adjust `spacelinestext`
The value of `spacelinestext` (distance from the lines to the baseline of the lyrics) was previously a fixed value, which meant that especially large lyrics would overlap with low notes (and small lyrics would seem very far from the lines).  By changing it to a font based value, the lyrics should naturally adjust their position based on their size.  The converted value corresponds to keeping the lyrics in the same place for 11pt Alegrya (the test font), as this causes the fewest tests to change.